refactor(waf): move AWS WAF challenge solver to flexget/utils/waf
The AWS WAF challenge-solving logic (AwsWaf, crypto, fingerprint spoofing,
PoW verify) was already protocol-generic with no IMDB coupling, so move it
out of flexget/components/imdb/awswaf/ into flexget/utils/waf/ for reuse by
future plugins. flexget/components/imdb/waf.py keeps IMDB-specific bits
(user-agent, headers, cookie domain, session singleton) as a thin wrapper
around the new generic waf_get()/solve_waf()/is_waf_challenge() API.

Also drops dead binance.com leftovers and a debug __main__ block picked up
from the vendored upstream source, and dedupes duplicated header dicts.
