--- flexget/components/imdb/awswaf/__init__.py
@@ -1,3 +1,0 @@
-from flexget.components.imdb.awswaf.aws import AwsWaf
-
-__all__ = ['AwsWaf']
