--- flexget/components/imdb/awswaf/crypto.py
@@ -1,29 +1,0 @@
-import os
-import base64
-from cryptography.hazmat.primitives.ciphers.aead import AESGCM
-
-key = bytes.fromhex("6f71a512b1e035eaab53d8be73120d3fb68a0ca346b9560aab3e5cdf753d5e98")
-aesgcm = AESGCM(key)
-
-def encrypt(plaintext: bytes) -> str:
-    iv = os.urandom(12)
-
-    cipher_bytes = aesgcm.encrypt(iv, plaintext, None)
-    tag = cipher_bytes[-16:]
-    ciphertext = cipher_bytes[:-16]
-
-    iv_b64 = base64.b64encode(iv).decode('utf-8')
-
-    return f"{iv_b64}::{tag.hex()}::{ciphertext.hex()}"
-
-def decrypt(encrypted: str) -> bytes:
-    iv_b64, tag_hex, ct_hex = encrypted.split("::")
-    iv = base64.b64decode(iv_b64)
-
-    tag = bytes.fromhex(tag_hex)
-    ciphertext = bytes.fromhex(ct_hex)
-
-    return aesgcm.decrypt(iv, ciphertext + tag, None)
-
-if __name__ == "__main__":
-    print(decrypt("DKDYyq/nvxsl4/Rs::7dd93738def231bad5a7166b860feb8e::d5f9c7dff9c626e02714b4407a6e81ac3f5d97c7b7ad43011e99982f08a4786082993b969fe9bcc6b8cf1876538ea9cc51c51ff356061e7714c211b56b8e9c8aef4d84cb543e85449264f3e2b19e3558e6952821940acc90aaf469f77b268d71fb99329ccbd4e80f34b98ce99a7d348d919324044fcb63981f06b4bafdc964cf1933f9f0fe88ad5f33956a9416f9b5b89df31ddec9a1f8bc9e2c91dac81172ecd94b0b1b1a6c2457d14d9c4e1207a984ae5830ed19381f77cd5019d4688557a9ab912b478e6bf28f1849b97a6ac2a24b64b94647926ddcb87af1ae5798f6413f2cb007a95b3e42b1ce9b7cec9280b561f9dc84e8cd6a2c0cb9d22e734ec26a0b010b2f06023a96f134961dac7178361c4bf4d9d1cbad17b6935daf14fce64c4702bcb359c31f8026acbba4ee707b7c34528821e2af219e5d234861694911e2fdb3666b1b6e81f6d56834fe49a8c7a8e0b5a55bdaf445d256459c63bc1d7769375440fa6e1d6ffbb8fa1865c78b6eb39125851f25f9e5712f75b73d62027bcf7a6608120079a61858cbd550275cfa938d521a9465fe98c7114ac25ae12397b47e914d2ff38c8141c1e25036ffcc8c58a58932c507e4749726affdb82f588cb1a6f30aeff0c1012d10bc913e159393badc29e8902a14c75824ef001e593682801e4f53946d89162aace235e6713ceb4a586eb140fb84d93342f0780c7a42a7f722c7e04afa7ce90feb49adf946cfd58a357047997b4dd9fa71341f8b13a3c92310c1bb1d64be06b3a4c4a7c7b330f4e9b689cc6a8990b6298a1e1a6a46a13a925028a72dde2af28ac7966143c312a579f2036030587049c64f505b897df94e1fef522d7b52a6d4505098fa5592a491f5753c227afe359badd370b1c75b7b179821dc1c03cd6df0153fb34d78df6c1a325b2d0015c028988e408b407c7c0e2015b63410fb66f68a0d11362f5e5a2bb5c74e3714869eddaf596fc9b62f6b4b8b800addf15fd8c2c606da5c9b9344a98591cd10e5980a4fc778487d347860c075e72f4ee4d3e9d105c07c2735f34ea8124c8ef4931ebde01c06e72b850107a033055261b8fb9ae1730b9fa02af58c032fde2b73b6f9c1cdc94b7ff7eb1d299c307133b4458622cad2a3c7e2d0e2d115cedf9eea1be423b610ebf7569a5908953c1ab377dc41312a180e6923e659d828e944f94c45c598b568f9fa7114a7ff038516ee1fde5b82417141980f459576006c45a8c5f535c770560e9cade747b44c319404766222139ad7db9228204a0416966e6f61a6b0037f9d8c045c4d556416032a7d2c92818ca0ec3a6f1f537e2a50851bb399582bb9e7bfeb9ff5f1a086dd5803e88805c75f63bdb633c2f63f5ecd5c6e2907e6210bffe73c8c6e4d6a57a2b82639dac4a126ffa1b8938856c06700b66f0083feb3aebf9efae9765a38e8250b898e1438db1fbbeda8de06dfb62647ebc1d18cead78080e6e984c05cec46068ea9982dad540bce166f6b95bf00735e8596d7f9ba61be4c47f921b557d8151ef3d7b8e7f5020fb89e3923e503f9922396ae6a0e5cbb2e5e3e6e0c5c08cd68ca0528770a908b2466c0efe2e82c44dfc1dd54df7b3379333f54046a0ed5467a01ebca2549dd1890c3676d9fe09e5c3c4f1bd4a6168e87e9b45f9d211aba9a414735ff698ba4a61ab7323a9b319f4e1ed1ade669b89d8960cdad96506c8218ec98ab6d34997b12dc546f7e98cbe7c77ea322d2db4cd689c87d23d62a3ff0a86e2780b5a4584292bf6e51704c617ba5260f6215e7f40b2683ec284eb4f16a411f9c6c2c9864f92749c7949f83ce12ccf35923d21194c697de57d29968524c8431f4e677812b1a7c59fe85dbfd31f63e68a73c1d02ad25adcfbdc565629ca74642c23475ff680beb20087cb3a10e87fd7618c24db2380162cc3863ba50230cad66a7bc0d119588ced9e6891e6d9277cc619feb92e138eb7bb86f05404d3aed4dea442fcf1ac1a455a340ee2cf86969b0164a34f3ed66b24cf390484924e8bf73c050c5c2a809248be0c91ac0684ab3fb3a31b4d71157b398cac0c2023ce37df73469a586dfafab3ff40547f9b4b9d2d250336bd5cc378b3d73f213d64af059046d6703d0c5dcf85e0ec4508b24eab64d325905753d222b04cb1385146f543ad6fa202c22f289d614967970ac2c414f10de1923e32552836b84c752a11483416e03061204b81adcc1d13cd413e155409403ad6e79fe96543c6f8fa3537a7626e609306cf979c20afa11c1d9e04c2826fa7c7be1bfeacf6345b1d11e9bc0136391f4bc7c98376ebc72dbbad65ffe17670d38270405c90895a867f663a539c1353e9c15567fa25ea662a56e5b442fc59bcf9a24ecfc456b476dba4fd41109d4cbc410745d543b5724ba066ee494384bcccb6159c54ab9b73f608129af434180ef583e2cb4efdaa2f73523897ecc0bc22039da78dc634d16e0aff8956a6c0d8d56a4c89a2b518b86bc5cc9a972c4c13ce14b9dfeb8ac54e0995f042a508038b9edea748eca2638ee467486523928791af0423ea7e1eb94bba592e9cc097b4d995b22c94e5585ed995e50df0adfd028c588489a902ae34625040eb69c1675ecb925117509c3368eb66e22d29ec30c157bac4d0a2563a86afe3cea70680bb186206eba5f35684661bd1c2aa04d3680d7e3cd55b59e7f654716cf8235f3fcf9406dacf907ac5b57f1867a3163112e14d1e98ebefd9d92f20f67a5801fd3ae096f8d7e1552c29cbd78ffb67f390df7760d9135c495536c3fa8f3233d91302f96b5aebf7a9363c43960861215cb836cd23a397a8d0f1d674db28e665d1f9331e3f853e9a5ee1c296d22761903c71c91e2dc06b3330dec97f5d563891739cfd187eedffc6614a58af063b0e123f18c92ef9d0d15e1aad7445bf1d23b2ba098df00be20f3901ab3f061cd903c5f57d78bafddca22b07e865d909221f90b67a37bd5e38b5e2ac3ac696771fc25f1c123a05448ec6499bd59191214c2dd1de488e065164b7b2422c0582668be91259306ea2b412aee768b90c767dc39cd2d23ae62fe453fe03be3c36e3a47c5fc388783b5fbb3160819f9480d4930735fffc602557a58240e3d7e422a30a0a41e7404c9e73effaa46961fcde795303b00ba7f6f572e7a62bedc3bd48bbcec7346e6f0fd5700d34b512406bf57d4619bd500e4bda73dcce550d59aa3c66ca686682cf9095bbb429ce834fea585ee1892e24950b7f733d300f82236e4b14cbaaa1bda7817aff9ae8a0fd9667c533eb35190445fa1230eeff7559446efaf9a9b0a8ba2b3b467a14c352469059354b0925b5bbefbe5bdaaeaa9a2c1ebbcc1ae6dceca1825625881f12365d86013cb8961f6aa880a62e8b1e5194c24e56e9bede5018e241e91dae920184281c9a094da4fdebcbdb8dc9c496eed566d97b2d89cd2ebc0bcf6278fc0b162fc3a5a0866ed412a6ea16b3f19f10acaf6c70086a6818ab670250bd3a627fc4c69c0da89805268b981c0250d5e540edf27b4ff456396ea91470e2907df9eff0100fa499bfe749e32f7a9c9c6fe5d1ecf89b98dacf7ddf4929fb898d4be30a0eaee59c37ff6202e5b3221c0c939a80a8b6762fb6417302b47ef8a56f10e7a9a2bbe6df420217c3194e5fcd49c75d1a17abea29b08ba52de2f5f8254b9373bc2d3a1547e4bdc5e874cf4228f2aa36801a8c84b09e5f45216bc7bc6ce27d94b483a56e329973cddae82cd8f874c133026c2c0830d1203d91da83270a6ed260d679136e60470df6275e913f28102c60c1f1f93cb76b2b73837cb9169e7045985ca43ef7a5d459f915d4d2354f47135396be1e9c2d12fdbf56cc8d2d4d19095aa49777bf134d46e01fb53d9e92b9513e43abbce4c68eb22fd53b37787a611b27f08bdecc88a95f4a8b8256f9409e42b52146f48d5b5a5105776ba1c60ba1df258f791678f898f20701c81d80b3f8e829bc36cb47d6220146ac238c11b3e057dfa1a9d4ef05ab9515e6397148a0f9d980fb340be117afeb89965cc4a90fcea5d5a16adbfebaf86c84d40e8c99661e90c33eaf3d062e2ca7d5c31fa39d23a2dc319fb91b0237d22b701e0e107e68a28bc233f3ce1e063c44e3bc0187893bfb274887672277839ccaa8d45d559e5e6aa47f91af262bc3c6c28d6440e83b3a212f44569ff85a647f50732c1ef707d2812466cdc6196f4cc70ce33382ea96370780d590d5edcbe02d1d500a79dd50f64d76a080e314cdaa54280410d3e350445d1583800bda2b7a9e774f7a08fad18b67f7e6e3c874f12c0a0b11488b5290701c9798e32810e11257e7dca2a06f8b7fa116e2d8cc783d856cc2e9996509ddb3234c3efc69a52b6841d6e8add5c3f5d87fd907b8cea986d4945531aca85529c7e7d4a4df633a08df231c7c60c691f2b0dca900b37a2a7f66e78bf46d09daa12bea74344edadda62217f58d1bf5c13e4ab2bd0f32187155f90c4c08ac74a40495426fff2505897c83239d962882916161e952506def4eb3c6726f8d046493f92175eb1e6bf6ada038f49d74921fc8a59131ea0a48c20e55751e087b458b1c31dc83eccfc3ae14e85e058edc0d54e50a2c29eead4266ef7f0c71278f3c989d0a326bab0efae795fcf0f4bcbdc0803bfb0d22cb76a7ee7a6f21f7a48ee3858fbfbaf6a101c0a6553c5dd8a2bc75a15411e5c103c62a61f6e5064679610818d861cdd2a6f6a808727f23abf2abf04ea56ecba34b55e66c300a76dc5803243271d850de6ef947c7abe19e9dedab5ea0c660ab8b7ddb61c010a8b4f20a614a31394c873184b39ae1d3c70addce834861f38200b1248f6dc0b842df3cb8a0e1d58b57ef6e50cfbe1642bf384da5d630f9f7007cf05ae1e1aeb176937e57d80b41b9ecc8a9c5ba0a1c000f4bd16c2eb32e56308e250bc4b0bde08f422c046a8773d2545cd610cce886e390daf1d8478e1b35208a3cfaf39b8ae6e56d2962c073c0e24623fb819adaf5c26ca3c1f51821320cb60219d05800b9c99abf89ba33f24aa9595bd1494092896e884a62008b5ac0ba9f32f396f1a803fda0dbf4c0a19f45fe4dce5e14090f06f49c48993e51f1e2a90adcbbe2d612212b238ca00cb48bd44db22f1f3e859"))
