+++ flexget/components/imdb/waf.py
@@ -3,9 +3,11 @@
 from __future__ import annotations
 
 import threading
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from loguru import logger
+
+from flexget.utils.waf import waf_get
 
 if TYPE_CHECKING:
     from curl_cffi.requests import Response
@@ -54,40 +56,15 @@ def set_domain_limiter(limiter) -> None:
     _domain_limiter = limiter
 
 
-def is_waf_challenge(response: Any) -> bool:
-    if response.status_code == 202:
-        return True
-    if response.headers.get('x-amzn-waf-action') == 'challenge':
-        return True
-    return bool(response.text and 'window.gokuProps' in response.text)
-
-
-def solve_waf(response: Any, domain: str = 'imdb.com') -> str:
-    from flexget.components.imdb.awswaf.aws import AwsWaf
-
-    goku, host = AwsWaf.extract(response.text)
-    logger.debug('Solving AWS WAF challenge for {}', domain)
-    token = AwsWaf(goku, host, domain, IMDB_USER_AGENT)()
-    logger.debug('AWS WAF challenge solved for {}', domain)
-    return token
-
-
 def imdb_get(url: str, *, raise_status: bool = True, **kwargs) -> Response:
     """GET an IMDB page, solving AWS WAF challenges when needed."""
-    if _domain_limiter is not None:
-        _domain_limiter()
-
-    session = _get_session()
-    headers = kwargs.pop('headers', None)
-    req_headers = {**session.headers, **headers} if headers else None
-
-    response = session.get(url, headers=req_headers, **kwargs)
-    if is_waf_challenge(response):
-        token = solve_waf(response)
-        session.cookies.set('aws-waf-token', token, domain='.imdb.com')
-        response = session.get(url, headers=req_headers, **kwargs)
-
-    if raise_status and response.status_code >= 400:
-        response.raise_for_status()
-
-    return response
+    return waf_get(
+        _get_session(),
+        url,
+        domain='imdb.com',
+        cookie_domain='.imdb.com',
+        user_agent=IMDB_USER_AGENT,
+        limiter=_domain_limiter,
+        raise_status=raise_status,
+        **kwargs,
+    )
