+++ flexget/utils/waf/verify.py
